First try to improve gestures again
