+++ KMaP/src/commonMain/kotlin/com/rafambn/kmap/gestures/GestureStateEnums.kt
@@ -19,5 +19,6 @@ enum class MapGestureState {
 enum class PathGestureState {
     WAITING_UP,
     WAITING_DOWN,
+    WAITING_DOWN_AFTER_TAP,
     WAITING_UP_AFTER_TAP,
 }
